+++ src/ReflectionTrait.php
@@ -40,7 +40,8 @@ trait ReflectionTrait
      * Gets all properties as array with property names as keys.
      *
      * If the object is a mock, it removes the properties added by PHPUnit.
-     * @param object $object Object from which to get properties
+     * @param string|object $object Instantiated object from which to get
+     *  properties or class name
      * @param int $filter The optional filter, for filtering desired property
      *  types. It's configured using `ReflectionProperty` constants, and
      *  default is public, protected and private properties
@@ -48,8 +49,9 @@ trait ReflectionTrait
      * @link http://php.net/manual/en/class.reflectionproperty.php#reflectionproperty.constants.modifiers
      * @since 1.1.4
      */
-    protected function getProperties(&$object, $filter = 256 | 512 | 1024)
+    protected function getProperties($object, $filter = 256 | 512 | 1024)
     {
+        $object = is_object($object) ? $object : new $object();
         $properties = (new ReflectionClass($object))->getProperties($filter);
 
         //Removes properties added by PHPUnit, if the object is a mock
@@ -82,26 +84,32 @@ trait ReflectionTrait
 
     /**
      * Gets a property value
-     * @param object $object Instantiated object that has the property
+     * @param string|object $object Instantiated object that has the property
+     *  or class name
      * @param string $name Property name
      * @return mixed Property value
      * @uses getPropertyInstance()
      */
-    protected function getProperty(&$object, $name)
+    protected function getProperty($object, $name)
     {
+        $object = is_object($object) ? $object : new $object();
+
         return $this->getPropertyInstance($object, $name)->getValue($object);
     }
 
     /**
      * Invokes a method
-     * @param object $object Instantiated object that we will run method on
+     * @param string|object $object Instantiated object that we will run method
+     *  on or class name
      * @param string $methodName Method name
      * @param array $parameters Array of parameters to pass into method
      * @return mixed Method return
      * @uses getMethodInstance()
      */
-    protected function invokeMethod(&$object, $methodName, array $parameters = [])
+    protected function invokeMethod($object, $methodName, array $parameters = [])
     {
+        $object = is_object($object) ? $object : new $object();
+
         return $this->getMethodInstance($object, $methodName)->invokeArgs($object, $parameters);
     }