+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # 1.x branch
 ## 1.2 branch
+### 1.2.7
+* `getProperty()`, `getProperties()` and `invokeMethod()` methods from the
+    `ReflectionTrait` can now take a class name as string, rather than just an
+    instantiated object.
+
 ### 1.2.6
 * fixed bug, the `TestCase::tearDown()` method removes temporary files only if
     these are not temporary system files;